[Issue#208][NT] Salma finished Google API
