--- arability/app/helpers/services_helper.rb
@@ -1,2 +1,0 @@
-module ServicesHelper
-end
