+++ arability/app/models/authentication.rb
@@ -1,3 +1,3 @@
 class Authentication < ActiveRecord::Base
-  attr_accessible :gid, :provider, :token, :token_secret
+  attr_accessible :gid, :gprovider, :token, :token_secret
 end
