--- arability/app/models/service.rb
@@ -1,4 +1,0 @@
-class Service < ActiveRecord::Base
-  belongs_to :gamer
-  attr_accessible :provider, :uid, :uname, :uemail
-end
