+++ arability/app/views/devise/registrations/new.html.erb
@@ -5,12 +5,6 @@
   <% if resource.errors.any? %>
     <div class="alert"><%= devise_error_messages! %></div>
   <% end %>
-
-    <div class="container">
-        <table style="width:900pt">
-        <col style="width:75%">
-        <col style="width:25%">
-        <tr><td>
     <div><%= f.label :email, :class => "control-label" %>
     <%= f.email_field :email, :style => "margin-left: 10px;" %></div></br>
 
@@ -39,15 +33,5 @@
     <div><%= f.submit "Sign up", :class => "btn btn-success" %></div>
 <% end %>
  </div>
-</td>
- <td style="width:25%; vertical-align:top">
-    <div class="container well" style="width:200px;background-color:white">
-         <div id="remote2"> 
-        </div>
-</td></tr>
-</table>
-</div>
-</div>
-</div>
 
 <%= render "links" %>
