--- arability/db/migrate/20130406064640_create_services.rb
@@ -1,13 +1,0 @@
-class CreateServices < ActiveRecord::Migration
-  def change
-    create_table :services do |t|
-      t.integer :user_id
-      t.string :provider
-      t.string :uid
-      t.string :uname
-      t.string :email
-
-      t.timestamps
-    end
-  end
-end
